db cleared

+++ app/src/main/java/com/example/sindoq/BlockAppsActivity.java
@@ -112,7 +112,7 @@ public class BlockAppsActivity extends Activity {
 
             for (int i = 0; i < appListMainArrayList.size(); i++) {
                 app = appListMainArrayList.get(i);
-                if(!"Sindoq".equals(app.getAppName().toString())) {
+                if(!"Sindoq".equals(app.getAppName().toString()) && !"Settings".equals(app.getAppName().toString())) {
                     if (databaseHelper.CHeckIfAppExistsInUnblocked(app.getAppName().toString()) == false) {
                         if (!databaseHelper.CHeckIfAppExists(app.getAppName().toString())) {
                             databaseHelper.insertapp(app.getAppName().toString(), app.getAppPackage().toString());
@@ -129,7 +129,7 @@ public class BlockAppsActivity extends Activity {
             public void onItemClick(View view, int position) {
                 appListMain = appListMainArrayList.get(position);
                 if (appListMain != null) {
-                    if(!"Sindoq".equals(appListMain.getAppName().toString())) {
+                    if(!"Sindoq".equals(appListMain.getAppName().toString())  && !"Settings".equals(appListMain.getAppName().toString() )) {
                         if (databaseHelper.CHeckIfAppExists(appListMain.getAppName().toString())) {
                             appListMain.setAppSelected(true);
                             databaseHelper.deleteApp(appListMain.getAppName().toString()); //delete from bloc); //Add in Unblocked Apps
